requirements.txt
streamlit==1.33.0
tensorflow==2.10.0
numpy==1.23.5
pillow